User create, 프로젝트 구조변경, 입력값 검증, 중복이메일 검사

+++ src/main/java/com/dhk/expensetrackerapi/user/controller/dto/request/UserRequest.java
@@ -23,6 +23,6 @@ public class UserRequest {
     @NotBlank(message = "User password must not be empty.")
     private String password;
     @NotNull(message = "User age must not be null")
-    private long age;
+    private int age;
 
 }
